+++ vector/src/main/java/im/vector/activity/VectorRoomInviteMembersActivity.java
@@ -22,6 +22,7 @@ import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.v4.app.FragmentActivity;
 import android.text.Editable;
 import android.text.TextUtils;
 import android.text.TextWatcher;
@@ -220,58 +221,14 @@ public class VectorRoomInviteMembersActivity extends VectorBaseSearchActivity {
                 boolean ret = false;
                 final Object item = mAdapter.getChild(groupPosition, childPosition);
 
-                if (item instanceof ParticipantAdapterItem){// && ((ParticipantAdapterItem) item).mIsValid) {
+                if (item instanceof ParticipantAdapterItem) {
                     final ParticipantAdapterItem participantAdapterItem = (ParticipantAdapterItem) item;
                     if (((ParticipantAdapterItem) item).mIsValid) {
                         finish(new ArrayList<>(Arrays.asList(participantAdapterItem)));
                         ret = true;
                     }
-                    else{
-                        if (ContactsManager.getInstance().isContactBookAccessAllowed()) {
-                            //enterEmailAddress(item.mContact);
-                            AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(VectorRoomInviteMembersActivity.this);
-                            alertDialogBuilder.setMessage(getString(R.string.people_invalid_warning_msg));
-                            final Context theContext = getApplicationContext();
-                            // set dialog message
-                            alertDialogBuilder
-                                    .setNegativeButton(R.string.cancel,null)
-                                    .setPositiveButton(R.string.action_edit_contact_form,
-                                            new DialogInterface.OnClickListener() {
-                                                public void onClick(DialogInterface dialog, int id) {
-                                                    DinsicUtils.editContactForm(theContext,VectorRoomInviteMembersActivity.this,getString(R.string.people_edit_contact_warning_msg),participantAdapterItem.mContact);
-                                                }
-                                            });
-
-                            // create alert dialog
-                            AlertDialog alertDialog = alertDialogBuilder.create();
-                            // show it
-                            alertDialog.show();
-
-                        }
-                        else {
-
-                            AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(VectorRoomInviteMembersActivity.this);
-                            alertDialogBuilder.setMessage(getString(R.string.people_invalid_warning_msg));
-
-                            // set dialog message
-                            alertDialogBuilder
-                                    .setCancelable(false)
-                                    .setPositiveButton(R.string.ok,
-                                            new DialogInterface.OnClickListener() {
-                                                public void onClick(DialogInterface dialog, int id) {
-
-                                                }
-                                            });
-
-                            // create alert dialog
-                            AlertDialog alertDialog = alertDialogBuilder.create();
-                            // show it
-                            alertDialog.show();
-
-                        }
-
-
-
+                    else {
+                        DinsicUtils.editContact(VectorRoomInviteMembersActivity.this,getApplicationContext(),(ParticipantAdapterItem) item);
                     }
                 }
                 return ret;
